# AnalyzerDriver performance improvements