… src/Compilers/Core/Portable/DiagnosticAnalyzer/AnalyzerDriver.GroupedAnalyzerActionsForAnalyzer.cs
… src/Compilers/Core/Portable/DiagnosticAnalyzer/AnalyzerDriver.cs
… src/Compilers/Core/Portable/DiagnosticAnalyzer/AnalyzerExecutor.cs
… src/Compilers/Core/Portable/DiagnosticAnalyzer/CompilationWithAnalyzers.cs
… src/Compilers/Core/Portable/DiagnosticAnalyzer/DiagnosticStartAnalysisScope.cs