@mavasani 2020-07-02
Ping @dotnet/roslyn-compiler for reviews