@mavasani 2020-07-06
Ping @dotnet/roslyn-compiler for reviews